FlagButton need to use UseEffect and cause Error
 Fixes:
ERROR    Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.     in Unknown (created by FlagButton)     in RCTView (at View.js:34)     in View (created by FlagButton)     in RCTView (at View.js:34)